[backend/frontend] some fixes
