+++ opencti-platform/opencti-front/src/components/dashboard/WidgetListCoreObjects.tsx
@@ -25,6 +25,7 @@ const bodyItemStyle = (width: string): CSSProperties => ({
 });
 
 interface WidgetListCoreObjectsProps {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   data: any[]
   dateAttribute: string
 }
