+++ opencti-platform/opencti-front/src/components/dashboard/WidgetMultiHeatMap.tsx
@@ -35,6 +35,7 @@ const lightColors = [
 ];
 
 interface WidgetMultiHeatMapProps {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   data: any[]
   minValue: number
   maxValue: number
