+++ opencti-platform/opencti-front/src/public/components/dashboard/PublicWidgetContainerProps.d.ts
