+++ opencti-platform/opencti-front/src/public/components/dashboard/stix_core_objects/PublicStixCoreObjectsHorizontalBars.tsx
@@ -1,17 +1,17 @@
 import { graphql, PreloadedQuery, usePreloadedQuery } from 'react-relay';
 import React from 'react';
 import { useTheme } from '@mui/styles';
-import type { PublicManifestWidget } from './PublicManifest';
-import { useFormatter } from '../../../components/i18n';
-import WidgetNoData from '../../../components/dashboard/WidgetNoData';
-import { itemColor } from '../../../utils/Colors';
-import { defaultValue } from '../../../utils/Graph';
-import WidgetHorizontalBars from '../../../components/dashboard/WidgetHorizontalBars';
-import type { Theme } from '../../../components/Theme';
-import type { PublicWidgetContainerProps } from './publicWidgetContainerProps';
-import useQueryLoading from '../../../utils/hooks/useQueryLoading';
-import WidgetContainer from '../../../components/dashboard/WidgetContainer';
-import WidgetLoader from '../../../components/dashboard/WidgetLoader';
+import type { PublicManifestWidget } from '../PublicManifest';
+import { useFormatter } from '../../../../components/i18n';
+import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
+import { itemColor } from '../../../../utils/Colors';
+import { defaultValue } from '../../../../utils/Graph';
+import WidgetHorizontalBars from '../../../../components/dashboard/WidgetHorizontalBars';
+import type { Theme } from '../../../../components/Theme';
+import type { PublicWidgetContainerProps } from '../PublicWidgetContainerProps';
+import useQueryLoading from '../../../../utils/hooks/useQueryLoading';
+import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
+import WidgetLoader from '../../../../components/dashboard/WidgetLoader';
 import { PublicStixCoreObjectsHorizontalBarsQuery } from './__generated__/PublicStixCoreObjectsHorizontalBarsQuery.graphql';
 
 const publicStixCoreObjectsHorizontalBarsQuery = graphql`
