+++ opencti-platform/opencti-front/src/public/components/dashboard/stix_core_objects/PublicStixCoreObjectsMultiHeatMap.tsx
@@ -1,15 +1,15 @@
 import { graphql, PreloadedQuery, usePreloadedQuery } from 'react-relay';
 import React from 'react';
-import WidgetNoData from '../../../components/dashboard/WidgetNoData';
-import type { PublicWidgetContainerProps } from './publicWidgetContainerProps';
-import { useFormatter } from '../../../components/i18n';
-import useQueryLoading from '../../../utils/hooks/useQueryLoading';
-import WidgetContainer from '../../../components/dashboard/WidgetContainer';
-import WidgetLoader from '../../../components/dashboard/WidgetLoader';
+import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
+import type { PublicWidgetContainerProps } from '../PublicWidgetContainerProps';
+import { useFormatter } from '../../../../components/i18n';
+import useQueryLoading from '../../../../utils/hooks/useQueryLoading';
+import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
+import WidgetLoader from '../../../../components/dashboard/WidgetLoader';
 import { PublicStixCoreObjectsMultiHeatMapQuery } from './__generated__/PublicStixCoreObjectsMultiHeatMapQuery.graphql';
-import WidgetMultiHeatMap from '../../../components/dashboard/WidgetMultiHeatMap';
-import type { PublicManifestWidget } from './PublicManifest';
-import { monthsAgo, now } from '../../../utils/Time';
+import WidgetMultiHeatMap from '../../../../components/dashboard/WidgetMultiHeatMap';
+import type { PublicManifestWidget } from '../PublicManifest';
+import { monthsAgo, now } from '../../../../utils/Time';
 
 const publicStixCoreObjectsMultiHeatMapQuery = graphql`
   query PublicStixCoreObjectsMultiHeatMapQuery(
