+++ opencti-platform/opencti-front/src/public/components/dashboard/stix_core_objects/PublicStixCoreObjectsNumber.tsx
@@ -1,13 +1,13 @@
 import { graphql, PreloadedQuery, usePreloadedQuery } from 'react-relay';
 import React from 'react';
-import useQueryLoading from '../../../utils/hooks/useQueryLoading';
+import useQueryLoading from '../../../../utils/hooks/useQueryLoading';
 import { PublicStixCoreObjectsNumberQuery } from './__generated__/PublicStixCoreObjectsNumberQuery.graphql';
-import WidgetContainer from '../../../components/dashboard/WidgetContainer';
-import { useFormatter } from '../../../components/i18n';
-import WidgetLoader from '../../../components/dashboard/WidgetLoader';
-import WidgetNumber from '../../../components/dashboard/WidgetNumber';
-import WidgetNoData from '../../../components/dashboard/WidgetNoData';
-import type { PublicWidgetContainerProps } from './publicWidgetContainerProps';
+import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
+import { useFormatter } from '../../../../components/i18n';
+import WidgetLoader from '../../../../components/dashboard/WidgetLoader';
+import WidgetNumber from '../../../../components/dashboard/WidgetNumber';
+import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
+import type { PublicWidgetContainerProps } from '../PublicWidgetContainerProps';
 
 const publicStixCoreObjectsNumberQuery = graphql`
   query PublicStixCoreObjectsNumberQuery(
