+++ opencti-platform/opencti-front/src/public/components/dashboard/stix_relationships/PublicStixRelationshipsHorizontalBars.tsx
@@ -1,17 +1,21 @@
 import { graphql, PreloadedQuery, usePreloadedQuery } from 'react-relay';
+import { useTheme } from '@mui/styles';
 import React from 'react';
-import type { PublicManifestWidget } from './PublicManifest';
-import WidgetDonut from '../../../components/dashboard/WidgetDonut';
-import WidgetNoData from '../../../components/dashboard/WidgetNoData';
-import type { PublicWidgetContainerProps } from './publicWidgetContainerProps';
-import { useFormatter } from '../../../components/i18n';
-import useQueryLoading from '../../../utils/hooks/useQueryLoading';
-import WidgetContainer from '../../../components/dashboard/WidgetContainer';
-import WidgetLoader from '../../../components/dashboard/WidgetLoader';
-import { PublicStixRelationshipsDonutQuery } from './__generated__/PublicStixRelationshipsDonutQuery.graphql';
+import type { PublicManifestWidget } from '../PublicManifest';
+import type { Theme } from '../../../../components/Theme';
+import { useFormatter } from '../../../../components/i18n';
+import { itemColor } from '../../../../utils/Colors';
+import { defaultValue } from '../../../../utils/Graph';
+import WidgetHorizontalBars from '../../../../components/dashboard/WidgetHorizontalBars';
+import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
+import type { PublicWidgetContainerProps } from '../PublicWidgetContainerProps';
+import useQueryLoading from '../../../../utils/hooks/useQueryLoading';
+import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
+import WidgetLoader from '../../../../components/dashboard/WidgetLoader';
+import { PublicStixRelationshipsHorizontalBarsQuery } from './__generated__/PublicStixRelationshipsHorizontalBarsQuery.graphql';
 
-const publicStixRelationshipsDonutQuery = graphql`
-  query PublicStixRelationshipsDonutQuery(
+const publicStixRelationshipsHorizontalBarsQuery = graphql`
+  query PublicStixRelationshipsHorizontalBarsQuery(
     $startDate: DateTime
     $endDate: DateTime
     $uriKey: String!
@@ -28,9 +32,11 @@ const publicStixRelationshipsDonutQuery = graphql`
       entity {
         ... on BasicObject {
           entity_type
+          id
         }
         ... on BasicRelationship {
           entity_type
+          id
         }
         ... on AttackPattern {
           name
@@ -125,15 +131,15 @@ const publicStixRelationshipsDonutQuery = graphql`
         }
         ... on DataComponent {
           name
-          description
         }
         ... on DataSource {
           name
-          description
         }
         ... on Case {
           name
-          description
+        }
+        ... on Report {
+          name
         }
         ... on StixCyberObservable {
           observable_value
@@ -141,6 +147,7 @@ const publicStixRelationshipsDonutQuery = graphql`
         ... on MarkingDefinition {
           definition_type
           definition
+          x_opencti_color
         }
         ... on KillChainPhase {
           kill_chain_name
@@ -171,17 +178,21 @@ const publicStixRelationshipsDonutQuery = graphql`
   }
 `;
 
-interface PublicStixRelationshipsDonutComponentProps {
+interface PublicStixRelationshipsHorizontalBarsComponentProps {
+  parameters: PublicManifestWidget['parameters']
   dataSelection: PublicManifestWidget['dataSelection']
-  queryRef: PreloadedQuery<PublicStixRelationshipsDonutQuery>
+  queryRef: PreloadedQuery<PublicStixRelationshipsHorizontalBarsQuery>
 }
 
-const PublicStixRelationshipsDonutComponent = ({
+const PublicStixRelationshipsHorizontalBarsComponent = ({
+  parameters,
   dataSelection,
   queryRef,
-}: PublicStixRelationshipsDonutComponentProps) => {
+}: PublicStixRelationshipsHorizontalBarsComponentProps) => {
+  const theme = useTheme<Theme>();
+  const { t_i18n } = useFormatter();
   const { publicStixRelationshipsDistribution } = usePreloadedQuery(
-    publicStixRelationshipsDonutQuery,
+    publicStixRelationshipsHorizontalBarsQuery,
     queryRef,
   );
 
@@ -189,19 +200,59 @@ const PublicStixRelationshipsDonutComponent = ({
     publicStixRelationshipsDistribution
     && publicStixRelationshipsDistribution.length > 0
   ) {
+    const selection = dataSelection[0];
+    const finalField = selection.attribute || 'entity_type';
+    const data = publicStixRelationshipsDistribution.map((n) => {
+      let color = selection.attribute?.endsWith('_id')
+        ? itemColor(n?.entity?.entity_type)
+        : itemColor(n?.label);
+      if (n?.entity?.color) {
+        color = theme.palette.mode === 'light' && n.entity.color === '#ffffff'
+          ? '#000000'
+          : n.entity.color;
+      }
+      if (n?.entity?.x_opencti_color) {
+        color = theme.palette.mode === 'light'
+        && n.entity.x_opencti_color === '#ffffff'
+          ? '#000000'
+          : n.entity.x_opencti_color;
+      }
+      return {
+        x: finalField.endsWith('_id')
+          ? defaultValue(n?.entity)
+          : n?.label,
+        y: n?.value,
+        fillColor: color,
+      };
+    });
+    const chartData = [{
+      name: selection.label || t_i18n('Number of relationships'),
+      data,
+    }];
+    const redirectionUtils = finalField.endsWith('_id')
+      ? publicStixRelationshipsDistribution.flatMap((n) => {
+        if (!n || !n.entity) return [];
+        return {
+          id: n.entity.id,
+          entity_type: n.entity.entity_type,
+        };
+      })
+      : undefined;
+
     return (
-      <WidgetDonut
-        data={[...publicStixRelationshipsDistribution]}
-        groupBy={dataSelection[0].attribute ?? 'entity_type'}
+      <WidgetHorizontalBars
+        series={chartData}
+        distributed={parameters.distributed}
         withExport={false}
         readonly={true}
+        redirectionUtils={redirectionUtils}
       />
     );
   }
   return <WidgetNoData />;
 };
 
-const PublicStixRelationshipsDonut = ({
+const PublicStixRelationshipsHorizontalBars = ({
   uriKey,
   widget,
   startDate,
@@ -210,8 +261,8 @@ const PublicStixRelationshipsDonut = ({
 }: PublicWidgetContainerProps) => {
   const { t_i18n } = useFormatter();
   const { id, parameters, dataSelection } = widget;
-  const queryRef = useQueryLoading<PublicStixRelationshipsDonutQuery>(
-    publicStixRelationshipsDonutQuery,
+  const queryRef = useQueryLoading<PublicStixRelationshipsHorizontalBarsQuery>(
+    publicStixRelationshipsHorizontalBarsQuery,
     {
       uriKey,
       widgetId: id,
@@ -222,13 +273,14 @@ const PublicStixRelationshipsDonut = ({
 
   return (
     <WidgetContainer
-      title={parameters.title ?? title ?? t_i18n('Entities number')}
+      title={parameters.title ?? title ?? t_i18n('Distribution of entities')}
       variant="inLine"
     >
       {queryRef ? (
         <React.Suspense fallback={<WidgetLoader />}>
-          <PublicStixRelationshipsDonutComponent
+          <PublicStixRelationshipsHorizontalBarsComponent
             queryRef={queryRef}
+            parameters={parameters}
             dataSelection={dataSelection}
           />
         </React.Suspense>
@@ -239,4 +291,4 @@ const PublicStixRelationshipsDonut = ({
   );
 };
 
-export default PublicStixRelationshipsDonut;
+export default PublicStixRelationshipsHorizontalBars;
