+++ opencti-platform/opencti-front/src/public/components/dashboard/stix_relationships/PublicStixRelationshipsMultiLineChart.tsx
@@ -1,15 +1,15 @@
 import { graphql, PreloadedQuery, usePreloadedQuery } from 'react-relay';
 import React from 'react';
-import type { PublicManifestWidget } from './PublicManifest';
-import { useFormatter } from '../../../components/i18n';
-import WidgetMultiLines from '../../../components/dashboard/WidgetMultiLines';
-import WidgetNoData from '../../../components/dashboard/WidgetNoData';
-import type { PublicWidgetContainerProps } from './publicWidgetContainerProps';
-import useQueryLoading from '../../../utils/hooks/useQueryLoading';
-import WidgetContainer from '../../../components/dashboard/WidgetContainer';
-import WidgetLoader from '../../../components/dashboard/WidgetLoader';
+import type { PublicManifestWidget } from '../PublicManifest';
+import { useFormatter } from '../../../../components/i18n';
+import WidgetMultiLines from '../../../../components/dashboard/WidgetMultiLines';
+import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
+import type { PublicWidgetContainerProps } from '../PublicWidgetContainerProps';
+import useQueryLoading from '../../../../utils/hooks/useQueryLoading';
+import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
+import WidgetLoader from '../../../../components/dashboard/WidgetLoader';
 import { PublicStixRelationshipsMultiLineChartQuery } from './__generated__/PublicStixRelationshipsMultiLineChartQuery.graphql';
-import { monthsAgo, now } from '../../../utils/Time';
+import { monthsAgo, now } from '../../../../utils/Time';
 
 const publicStixRelationshipsMultiLineChartQuery = graphql`
   query PublicStixRelationshipsMultiLineChartQuery(
