+++ opencti-platform/opencti-graphql/src/utils/format.js
@@ -74,7 +74,12 @@ export const computeRangeIntersection = (a, b) => {
 
 export const minutesAgo = (minutes) => moment().utc().subtract(minutes, 'minutes');
 export const hoursAgo = (hours) => moment().utc().subtract(hours, 'hours');
-export const dayAgo = (days = 1) => {
+
+/**
+ * @param {number} days Number of days
+ * @return {string} ISO Date string
+ */
+export const daysAgo = (days) => {
   const currentDate = new Date();
   currentDate.setDate(currentDate.getDate() - days);
   return currentDate.toISOString().split('T')[0];
